handle error on handleURLRedirect
